reorganize `vortex-scan` modules
